+++ app/react/portainer/environments/ItemView/UpdateForm/AgentForm/types.tsx
@@ -1,0 +1,11 @@
+export interface FormValues {
+  name: string;
+
+  url: string;
+  publicUrl: string;
+
+  meta: {
+    tagIds: number[];
+    groupId: number;
+  };
+}
