+++ app/react/portainer/environments/ItemView/UpdateForm/AzureForm/types.tsx
@@ -1,0 +1,11 @@
+import { EnvironmentMetadata } from '@/react/portainer/environments/environment.service/create';
+
+export interface FormValues {
+  name: string;
+
+  meta: EnvironmentMetadata;
+
+  applicationId: string;
+  tenantId: string;
+  authKey: string;
+}
