+++ app/react/portainer/environments/ItemView/UpdateForm/OtherTypesForm/types.tsx
@@ -1,0 +1,11 @@
+import { EnvironmentMetadata } from '@/react/portainer/environments/environment.service/create';
+
+import { TLSConfig } from '@@/TLSFieldset/types';
+
+export interface FormValues {
+  name: string;
+  url: string;
+  publicUrl: string;
+  tlsConfig: TLSConfig;
+  meta: EnvironmentMetadata;
+}
